% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bunnylatin.R
\name{bunnylatin}
\alias{bunnylatin}
\title{Bunny Latin}
\usage{
bunnylatin(x)
}
\arguments{
\item{x}{A string, or strings}
}
\value{
A modified string, according to Bunny Latin Rules
}
\description{
Bunny Latin
}
\details{
Bunny Latin will move the last two letters of the word to the front, then repeat the last letter of the newly modified word once at the end.
}
\examples{
bunnylatin('apple')

x <- c('summer', 'sunshine', 'puppy')
bunnylatin(x)

}
