% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bunnylatin.R
\name{bunnylatin}
\alias{bunnylatin}
\title{bunnylatin}
\usage{
bunnylatin(x)
}
\arguments{
\item{x}{A string or vector. Each string length must be equal to or greater than three.}
}
\value{
A modified string or vector, according to the Bunny Latin Rules.
}
\description{
\code{bunnylatin} will move the last two letters of the word to the front, then repeat the last letter of the newly modified word once at the end.
}
\examples{
bunnylatin('apple')

bunnylatin('12345')

x <- c('summer', 'sunshine', 'puppy')
bunnylatin(x)


}
